+++ sw/airborne/firmwares/rotorcraft/guidance/guidance_indi_hybrid.h
@@ -70,7 +70,7 @@ enum GuidanceIndiHybrid_VMode {
 
 extern struct StabilizationSetpoint guidance_indi_run(struct FloatVect3 *accep_sp, float heading_sp);
 extern struct StabilizationSetpoint guidance_indi_run_mode(bool in_flight, struct HorizontalGuidance *gh, struct VerticalGuidance *gv, enum GuidanceIndiHybrid_HMode h_mode, enum GuidanceIndiHybrid_VMode v_mode);
-extern void guidance_indi_set_min_max_airspeed(float min_airspeed, float max_airspeed);
+extern void guidance_set_min_max_airspeed(float min_airspeed, float max_airspeed);
 
 struct guidance_indi_hybrid_params {
   float pos_gain;
