+++ sw/airborne/modules/ctrl/eff_scheduling_rotwing.c
@@ -173,6 +173,8 @@ inline float bound_or_zero(float value, float low_lim, float up_lim) {
 float eff_sched_pusher_time = 0.002;
 float roll_eff_slider = 12.f;
 
+static const float Wu_gih_original[GUIDANCE_INDI_HYBRID_U] = GUIDANCE_INDI_WLS_WU;
+
 struct rotwing_eff_sched_var_t eff_sched_var;
 
 inline void eff_scheduling_rotwing_update_wing_angle(void);
@@ -189,6 +191,7 @@ inline void eff_scheduling_rotwing_schedule_liftd(void);
 
 inline float guidance_indi_get_liftd(float pitch UNUSED, float theta UNUSED);
 void stabilization_indi_set_wls_settings(void);
+inline void guidance_indi_hybrid_set_wls_settings(float body_v[3], float roll_angle, float pitch_angle);
 
 
 /** ABI binding wing position data.
@@ -495,4 +498,92 @@ void stabilization_indi_set_wls_settings(void)
         Bound(wls_stab_p.u_max[i], 0, MAX_PPRZ);
       }
   }
+}
+
+void guidance_indi_hybrid_set_wls_settings(float body_v[3], float roll_angle, float pitch_angle)
+{
+  // adjust weights
+  float fixed_wing_percentile = (rotwing_state_hover_motors_idling())? 1:0; // TODO: when hover props go below 40%, ...
+  Bound(fixed_wing_percentile, 0, 1);
+#define AIRSPEED_IMPORTANCE_IN_FORWARD_WEIGHT 16
+
+  float Wv_original[GUIDANCE_INDI_HYBRID_V] = GUIDANCE_INDI_WLS_PRIORITIES;
+
+  // Increase importance of forward acceleration in forward flight
+  wls_guid_p.Wv[0] = Wv_original[0] * (1.0f + fixed_wing_percentile *
+                                         AIRSPEED_IMPORTANCE_IN_FORWARD_WEIGHT); // stall n low hover motor_off (weight 16x more important than vertical weight)
+
+  struct FloatEulers eulers_zxy;
+  float_eulers_of_quat_zxy(&eulers_zxy, stateGetNedToBodyQuat_f());
+
+  float du_min_thrust_z = ((MAX_PPRZ - actuator_state_filt_vect[0]) * g1g2[3][0] + (MAX_PPRZ -
+                           actuator_state_filt_vect[1]) * g1g2[3][1] + (MAX_PPRZ - actuator_state_filt_vect[2]) * g1g2[3][2] +
+                           (MAX_PPRZ - actuator_state_filt_vect[3]) * g1g2[3][3]) * rotwing_state_hover_motors_running();
+  Bound(du_min_thrust_z, -50., 0.);
+  float du_max_thrust_z = -(actuator_state_filt_vect[0] * g1g2[3][0] + actuator_state_filt_vect[1] * g1g2[3][1] +
+                            actuator_state_filt_vect[2] * g1g2[3][2] + actuator_state_filt_vect[3] * g1g2[3][3]);
+  Bound(du_max_thrust_z, 0., 50.);
+
+  float roll_limit_rad = guidance_indi_max_bank;
+  float max_pitch_limit_rad = RadOfDeg(GUIDANCE_INDI_MAX_PITCH);
+  float min_pitch_limit_rad = RadOfDeg(GUIDANCE_INDI_MIN_PITCH);
+
+  float fwd_pitch_limit_rad = RadOfDeg(GUIDANCE_INDI_MAX_PITCH);
+  float quad_pitch_limit_rad = RadOfDeg(5.0);
+
+  float airspeed = stateGetAirspeed_f();
+
+  float scheduled_pitch_angle = 0.f;
+  float pitch_angle_range = 3.;
+  float meas_skew_angle = rotwing_state.meas_skew_angle_deg;
+  Bound(meas_skew_angle, 0, 90); // Bound to prevent errors
+  if (meas_skew_angle < ROTWING_SKEW_ANGLE_STEP) {
+    scheduled_pitch_angle = ROTWING_QUAD_PREF_PITCH;
+    wls_guid_p.Wu[1] = Wu_gih_original[1];
+    max_pitch_limit_rad = quad_pitch_limit_rad;
+  } else {
+    float pitch_progression = (airspeed - rotwing_state.fw_min_airspeed) / 2.f;
+    Bound(pitch_progression, 0.f, 1.f);
+    scheduled_pitch_angle = pitch_angle_range * pitch_progression + ROTWING_QUAD_PREF_PITCH*(1.f-pitch_progression);
+    wls_guid_p.Wu[1] = Wu_gih_original[1] * (1.f - pitch_progression*0.99);
+    max_pitch_limit_rad = quad_pitch_limit_rad + (fwd_pitch_limit_rad - quad_pitch_limit_rad) * pitch_progression;
+  }
+  if (!rotwing_state_hover_motors_running()) {
+    scheduled_pitch_angle = 8.;
+    max_pitch_limit_rad = fwd_pitch_limit_rad;
+  }
+  Bound(scheduled_pitch_angle, -5., 8.);
+  guidance_indi_pitch_pref_deg = scheduled_pitch_angle;
+
+  float pitch_pref_rad = RadOfDeg(guidance_indi_pitch_pref_deg);
+
+  // Set lower limits
+  wls_guid_p.u_min[0] = -roll_limit_rad - roll_angle; //roll
+  wls_guid_p.u_min[1] = min_pitch_limit_rad - pitch_angle; // pitch
+
+  // Set upper limits limits
+  wls_guid_p.u_max[0] = roll_limit_rad - roll_angle; //roll
+  wls_guid_p.u_max[1] = max_pitch_limit_rad - pitch_angle; // pitch
+
+  if(rotwing_state_hover_motors_running()) {
+    wls_guid_p.u_min[2] = du_min_thrust_z;
+    wls_guid_p.u_max[2] = du_max_thrust_z;
+  } else {
+    wls_guid_p.u_min[2] = 0.;
+    wls_guid_p.u_max[2] = 0.;
+  }
+
+  if(rotwing_state_pusher_motor_running()) {
+    wls_guid_p.u_min[3] = (-actuator_state_filt_vect[8] * g1g2[4][8]);
+    wls_guid_p.u_max[3] = 9.0; // Hacky value to prevent drone from pitching down in transition
+  } else {
+    wls_guid_p.u_min[3] = 0.;
+    wls_guid_p.u_max[3] = 0.;
+  }
+
+  // Set prefered states
+  wls_guid_p.u_pref[0] = 0; // prefered delta roll angle
+  wls_guid_p.u_pref[1] = -pitch_angle + pitch_pref_rad;// prefered delta pitch angle
+  wls_guid_p.u_pref[2] = wls_guid_p.u_max[2]; // Low thrust better for efficiency
+  wls_guid_p.u_pref[3] = body_v[0]; // solve the body acceleration
 }
