+++ sw/airborne/modules/rotwing_drone/rotwing_state.c
@@ -24,10 +24,8 @@
  */
 
 #include "modules/rotwing_drone/rotwing_state.h"
-#include "firmwares/rotorcraft/stabilization/stabilization_indi.h"
-#include "firmwares/rotorcraft/guidance/guidance_indi_hybrid.h"
 #include "firmwares/rotorcraft/autopilot_firmware.h"
-
+#include "modules/core/commands.h"
 #include "modules/actuators/actuators.h"
 #include "modules/core/abi.h"
 
@@ -61,11 +59,6 @@
 #define ROTWING_FW_SKEW_ANGLE 85.0
 #endif
 
-/* Magnitude skew angle jump away from quad */
-#ifndef ROTWING_SKEW_ANGLE_STEP
-#define ROTWING_SKEW_ANGLE_STEP 55.0
-#endif
-
 /* TODO: Give a name.... */
 #ifndef ROTWING_SKEW_BACK_MARGIN
 #define ROTWING_SKEW_BACK_MARGIN 5.0
@@ -79,11 +72,6 @@
 /* Skew angle at which the mininum airspeed starts its linear portion */
 #ifndef ROTWING_MIN_AIRSPEED_SLOPE_START_ANGLE
 #define ROTWING_MIN_AIRSPEED_SLOPE_START_ANGLE 30.0
-#endif
-
-/* Preferred pitch angle for the quad mode (deg) */
-#ifndef ROTWING_QUAD_PREF_PITCH
-#define ROTWING_QUAD_PREF_PITCH -5.0
 #endif
 
 /* Amount of time the airspeed needs to be below the FW_MIN_AIRSPEED */
@@ -107,13 +95,7 @@
 #endif
 abi_event rotwing_state_feedback_ev;
 static void rotwing_state_feedback_cb(uint8_t sender_id, struct act_feedback_t *feedback_msg, uint8_t num_act);
-
-static bool rotwing_state_hover_motors_idling(void);
-static const float Wu_gih_original[GUIDANCE_INDI_HYBRID_U] = GUIDANCE_INDI_WLS_WU;
 struct rotwing_state_t rotwing_state;
-
-inline void guidance_indi_hybrid_set_wls_settings(float body_v[3], float roll_angle, float pitch_angle);
-
 #if PERIODIC_TELEMETRY
 #include "modules/datalink/telemetry.h"
 static void send_rotating_wing_state(struct transport_tx *trans, struct link_device *dev)
@@ -164,7 +146,6 @@ void rotwing_state_init(void)
   rotwing_state.fail_hover_motor = false;
   rotwing_state.fail_pusher_motor = false;
   rotwing_state.ref_model_skew_angle_deg = 0;
-
   // Bind ABI messages
   AbiBindMsgACT_FEEDBACK(ROTWING_STATE_ACT_FEEDBACK_ID, &rotwing_state_feedback_ev, rotwing_state_feedback_cb);
 
@@ -177,7 +158,7 @@ void rotwing_state_init(void)
  * @brief Check if hover motors are idling (COMMAND_THRUST < ROTWING_QUAD_IDLE_MIN_THRUST) for ROTWING_QUAD_IDLE_TIMEOUT time
  * @return true if hover motors are idling, false otherwise
  */
-static bool rotwing_state_hover_motors_idling(void) {
+bool rotwing_state_hover_motors_idling(void) {
   static float last_idle_time = 0;
   // Check if hover motors are idling and reset timer
   if(stabilization.cmd[COMMAND_THRUST] > ROTWING_QUAD_IDLE_MIN_THRUST) {
@@ -306,9 +287,7 @@ void rotwing_state_periodic(void)
     rotwing_state.max_airspeed = 0; // Max airspeed FW
   }*/
 
-  guidance_indi_set_min_max_airspeed(rotwing_state.min_airspeed, rotwing_state.max_airspeed);
-
-
+  guidance_set_min_max_airspeed(rotwing_state.min_airspeed, rotwing_state.max_airspeed);
   /* Set navigation/guidance settings */
   nav_max_deceleration_sp = ROTWING_FW_MAX_DECELERATION * meas_skew_angle / 90.f + ROTWING_QUAD_MAX_DECELERATION * (90.f - meas_skew_angle) / 90.f; //TODO: Do we really want to based this on the skew?
 
@@ -333,13 +312,18 @@ void rotwing_state_periodic(void)
 #else
   rotwing_state.skew_cmd = servo_pprz_cmd;
 #endif
-
+#ifdef COMMAND_ROT_MECH
+  commands[COMMAND_ROT_MECH] = rotwing_state.skew_cmd;
+#endif
 
   /* Add simulation feedback for the skewing and RPM */
 #if USE_NPS
   // Export to the index of the SKEW in the NPS_ACTUATOR_NAMES array
+#ifdef COMMAND_ROT_MECH
+  commands[COMMAND_ROT_MECH] = (rotwing_state.skew_cmd + MAX_PPRZ) / 2.f; // Scale to simulation command
+#else
   actuators_pprz[INDI_NUM_ACT] = (rotwing_state.skew_cmd + MAX_PPRZ) / 2.f; // Scale to simulation command
-
+#endif
   // SEND ABI Message to ctr_eff_sched, ourself and other modules that want Actuator position feedback
   struct act_feedback_t feedback;
   feedback.idx =  SERVO_ROTATION_MECH_IDX;
@@ -446,94 +430,6 @@ bool rotwing_state_skew_angle_valid(void) {
   bool skew_angle_match = fabs(rotwing_state.meas_skew_angle_deg - rotwing_state.ref_model_skew_angle_deg) < ROTWING_SKEW_REF_MODEL_MAX_DIFF;
   
   return (!skew_timeout && skew_angle_match);
-}
-
-void guidance_indi_hybrid_set_wls_settings(float body_v[3], float roll_angle, float pitch_angle)
-{
-  // adjust weights
-  float fixed_wing_percentile = (rotwing_state_hover_motors_idling())? 1:0; // TODO: when hover props go below 40%, ...
-  Bound(fixed_wing_percentile, 0, 1);
-#define AIRSPEED_IMPORTANCE_IN_FORWARD_WEIGHT 16
-
-  float Wv_original[GUIDANCE_INDI_HYBRID_V] = GUIDANCE_INDI_WLS_PRIORITIES;
-
-  // Increase importance of forward acceleration in forward flight
-  wls_guid_p.Wv[0] = Wv_original[0] * (1.0f + fixed_wing_percentile *
-                                         AIRSPEED_IMPORTANCE_IN_FORWARD_WEIGHT); // stall n low hover motor_off (weight 16x more important than vertical weight)
-
-  struct FloatEulers eulers_zxy;
-  float_eulers_of_quat_zxy(&eulers_zxy, stateGetNedToBodyQuat_f());
-
-  float du_min_thrust_z = ((MAX_PPRZ - actuator_state_filt_vect[0]) * g1g2[3][0] + (MAX_PPRZ -
-                           actuator_state_filt_vect[1]) * g1g2[3][1] + (MAX_PPRZ - actuator_state_filt_vect[2]) * g1g2[3][2] +
-                           (MAX_PPRZ - actuator_state_filt_vect[3]) * g1g2[3][3]) * rotwing_state_hover_motors_running();
-  Bound(du_min_thrust_z, -50., 0.);
-  float du_max_thrust_z = -(actuator_state_filt_vect[0] * g1g2[3][0] + actuator_state_filt_vect[1] * g1g2[3][1] +
-                            actuator_state_filt_vect[2] * g1g2[3][2] + actuator_state_filt_vect[3] * g1g2[3][3]);
-  Bound(du_max_thrust_z, 0., 50.);
-
-  float roll_limit_rad = guidance_indi_max_bank;
-  float max_pitch_limit_rad = RadOfDeg(GUIDANCE_INDI_MAX_PITCH);
-  float min_pitch_limit_rad = RadOfDeg(GUIDANCE_INDI_MIN_PITCH);
-
-  float fwd_pitch_limit_rad = RadOfDeg(GUIDANCE_INDI_MAX_PITCH);
-  float quad_pitch_limit_rad = RadOfDeg(5.0);
-
-  float airspeed = stateGetAirspeed_f();
-
-  float scheduled_pitch_angle = 0.f;
-  float pitch_angle_range = 3.;
-  float meas_skew_angle = rotwing_state.meas_skew_angle_deg;
-  Bound(meas_skew_angle, 0, 90); // Bound to prevent errors
-  if (meas_skew_angle < ROTWING_SKEW_ANGLE_STEP) {
-    scheduled_pitch_angle = ROTWING_QUAD_PREF_PITCH;
-    wls_guid_p.Wu[1] = Wu_gih_original[1];
-    max_pitch_limit_rad = quad_pitch_limit_rad;
-  } else {
-    float pitch_progression = (airspeed - rotwing_state.fw_min_airspeed) / 2.f;
-    Bound(pitch_progression, 0.f, 1.f);
-    scheduled_pitch_angle = pitch_angle_range * pitch_progression + ROTWING_QUAD_PREF_PITCH*(1.f-pitch_progression);
-    wls_guid_p.Wu[1] = Wu_gih_original[1] * (1.f - pitch_progression*0.99);
-    max_pitch_limit_rad = quad_pitch_limit_rad + (fwd_pitch_limit_rad - quad_pitch_limit_rad) * pitch_progression;
-  }
-  if (!rotwing_state_hover_motors_running()) {
-    scheduled_pitch_angle = 8.;
-    max_pitch_limit_rad = fwd_pitch_limit_rad;
-  }
-  Bound(scheduled_pitch_angle, -5., 8.);
-  guidance_indi_pitch_pref_deg = scheduled_pitch_angle;
-
-  float pitch_pref_rad = RadOfDeg(guidance_indi_pitch_pref_deg);
-
-  // Set lower limits
-  wls_guid_p.u_min[0] = -roll_limit_rad - roll_angle; //roll
-  wls_guid_p.u_min[1] = min_pitch_limit_rad - pitch_angle; // pitch
-
-  // Set upper limits limits
-  wls_guid_p.u_max[0] = roll_limit_rad - roll_angle; //roll
-  wls_guid_p.u_max[1] = max_pitch_limit_rad - pitch_angle; // pitch
-
-  if(rotwing_state_hover_motors_running()) {
-    wls_guid_p.u_min[2] = du_min_thrust_z;
-    wls_guid_p.u_max[2] = du_max_thrust_z;
-  } else {
-    wls_guid_p.u_min[2] = 0.;
-    wls_guid_p.u_max[2] = 0.;
-  }
-
-  if(rotwing_state_pusher_motor_running()) {
-    wls_guid_p.u_min[3] = (-actuator_state_filt_vect[8] * g1g2[4][8]);
-    wls_guid_p.u_max[3] = 9.0; // Hacky value to prevent drone from pitching down in transition
-  } else {
-    wls_guid_p.u_min[3] = 0.;
-    wls_guid_p.u_max[3] = 0.;
-  }
-
-  // Set prefered states
-  wls_guid_p.u_pref[0] = 0; // prefered delta roll angle
-  wls_guid_p.u_pref[1] = -pitch_angle + pitch_pref_rad;// prefered delta pitch angle
-  wls_guid_p.u_pref[2] = wls_guid_p.u_max[2]; // Low thrust better for efficiency
-  wls_guid_p.u_pref[3] = body_v[0]; // solve the body acceleration
 }
 
 void rotwing_state_set(enum rotwing_states_t state) {
